Fix: Addressed NameError in data_operations.py. Refactor: Moved check_polygon_closure(), warn_if_polygon_not_closed(), and geopy_distance import to data_operations.py from util.py. Added: CHANGELOG.md.
