+++ src/utils.py
@@ -195,30 +195,6 @@ def parse_dd_or_dms(coordinate_format):
         print("Invalid format passed to the function.")
         return None
 
-            
-def is_polygon_close_to_being_closed(points, threshold=20):
-    for i in range(len(points) - 1):  # Exclude the last point itself
-        distance = geopy_distance(points[-1], points[i]).feet
-        if distance <= threshold:
-            return True
-    return False
-    
-
-def check_polygon_closure(points, reference_point=None):
-    if len(points) > 2:
-        distance_between_first_and_last = geopy_distance(points[0], points[-1]).feet
-        
-        # Check if the distance is extremely close
-        if distance_between_first_and_last < 0.1:
-            points[-1] = points[0]
-            return True
-        
-        if reference_point:
-            distance_from_reference_to_last = geopy_distance(reference_point[:2], points[-1]).feet
-            return distance_between_first_and_last <= 10 or distance_from_reference_to_last <= 10
-        else:
-            return distance_between_first_and_last <= 10
-    return False
 
 def transform_tzt_data_to_kml_format(parsed_data):
     data = {
